Show Advanced fields if data / gas limit is preset
It is possible to prefill transaction fields by setting query parameters on
the myetherwallet url. For example:

https://myetherwallet.com/interface/send-transaction?to=0xa0b86991c6218b36c1d19d4a2e9eb0ce3606eb48&value=0&data=0x095ea7b3000000000000000000000000ff1cafa239e3b123964519a3b2e37d808de905028000000000000000000000000000000000000000000000000000000000000000&gaslimit=250000&network=eth#send-transaction

However, if you prefill the data or gas limit fields, those fields are not visible on
the "Send Transaction" page because the fields are hidden in the "Advanced" section.

The data and gas limit fields should be visible on the "Send Transaction" page if they are prefilled by the url.