bitbox01: hotfix on v5: make BitBox01 work with WebHID, drop U2F-hijack
BitBox01 is currently not functional on v6.0.3 and it does not seem
trivial to make it work.

The previous version is hosted at https://v5.myetherwallet.com/, but
there, BitBox01 support also broke.

This patch is a hot-fix for the v5 deployment that makes the BitBox01
integration functional again, using WebHID (Chrome only).

This will help BitBox01 users migrate their coins to a new wallet.